% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myConstr.R
\name{myConstr}
\alias{myConstr}
\title{Connstrutor function for t tests}
\usage{
myConstr(x, y, alpha = 0.05)
}
\arguments{
\item{x}{vector of data}

\item{y}{vector of data}

\item{alpha}{alpha level}
}
\value{
a list containing the data frame and relevant information
}
\description{
Connstrutor function for t tests
}
\examples{
myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10))
}
