% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPrint.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print function for t tests for project}
\usage{
\method{print}{Rttest}(obj)
}
\value{
data frame containing x and y , p value, apha and confidence interval
}
\description{
Print function for t tests for project
}
\author{
Amina Asghar
}
