% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print function for t tests}
\usage{
\method{print}{Rttest}(obj)
}
\value{
data frame containing x and y , p value, apha and confidence interval
}
\description{
Print function for t tests
}
\author{
Amina Asghar
}
